requirements_macos.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
boto3==1.34.144
botocore==1.34.162
bs4==0.0.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.11
cryptography
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Dance==7.1.0
Flask-SocketIO==5.3.6
future==1.0.0
gevent==24.2.1
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
loguru==0.7.2
lxml==5.3.0
macholib==1.16.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
numpy==2.1.0
oa_atomacos==3.2.0
oauthlib==3.2.2
obsws-python==1.7.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.41
PyGetWindow==0.0.4
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
PyRect==0.2.0
PyScreeze==0.1.19
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytweening==1.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
rubicon-objc==0.4.9
s3transfer==0.10.2
screeninfo==0.8.1
setuptools==72.1.0
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.6
urllib3==2.2.2
URLObject==2.4.3
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.1
mss==9.0.1
plyer==2.1.0
uiautomation==2.0.20